+++ projects/fusion-ui/components/tooltip/v2/tooltip.directive.spec.ts
@@ -2,7 +2,7 @@ import {TooltipDirective} from './tooltip.directive';
 
 describe('TooltipDirective', () => {
     it('should create an instance', () => {
-        const directive = new TooltipDirective(null, null, null);
+        const directive = new TooltipDirective(null, null, null, null);
         expect(directive).toBeTruthy();
     });
 });